[Feature] #5 - Edit DataModel
