+++ SKHUAZ/SKHUAZ/Presentation/Recommend/DataModel/RecommendDM.swift
@@ -1,0 +1,36 @@
+//
+//  RecommendDetailDataModel.swift
+//  SKHUAZ
+//
+//  Created by 박신영 on 2023/08/20.
+//
+
+import Foundation
+
+
+struct RecommendDetailDataModel {
+    var nickNameLabel: String
+    var courseYearLabel: String
+    var titleLabel: String
+    var majorNameLabel: String
+    var creationDateLabel: String
+    var routeLabel: String
+    var contentLabel: String
+}
+
+let rdReview1 = RecommendDetailDataModel(nickNameLabel: "박신영", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 소프1", majorNameLabel: "소프트웨어공학과", creationDateLabel: "2023-08-19", routeLabel: "자바>웹입>등등", contentLabel: "이렇게만 듣자!")
+
+let rdReview2 = RecommendDetailDataModel(nickNameLabel: "천성우", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 소프2", majorNameLabel: "소프트웨어공학과", creationDateLabel: "2023-08-20", routeLabel: "자바>웹입>등등", contentLabel: "요렇게만 듣자!")
+
+let rdReview3 = RecommendDetailDataModel(nickNameLabel: "문인호", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 정통1", majorNameLabel: "정보통신공학과", creationDateLabel: "2023-08-19", routeLabel: "컴구, 데통>컴네>등등", contentLabel: "고렇게만 듣자!")
+
+let rdReview4 = RecommendDetailDataModel(nickNameLabel: "문인호", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 정통1", majorNameLabel: "정보통신공학과", creationDateLabel: "2023-08-19", routeLabel: "컴구, 데통>컴네>등등", contentLabel: "고렇게만 듣자!")
+
+let rdReview5 = RecommendDetailDataModel(nickNameLabel: "문인호", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 정통1", majorNameLabel: "정보통신공학과", creationDateLabel: "2023-08-19", routeLabel: "컴구, 데통>컴네>등등", contentLabel: "고렇게만 듣자!")
+
+let rdReview6 = RecommendDetailDataModel(nickNameLabel: "문인호", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 정통1", majorNameLabel: "정보통신공학과", creationDateLabel: "2023-08-19", routeLabel: "컴구, 데통>컴네>등등", contentLabel: "고렇게만 듣자!")
+
+let rdReview7 = RecommendDetailDataModel(nickNameLabel: "문인호", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 정통1", majorNameLabel: "정보통신공학과", creationDateLabel: "2023-08-19", routeLabel: "컴구, 데통>컴네>등등", contentLabel: "고렇게만 듣자!")
+
+//let rdReview4 = RecommendDetailDataModel(nickNameLabel: "누구세용", courseYearLabel: "2023-1", titleLabel: "드디어 공개한다 인공지능1", majorNameLabel: "인공지능공학과", creationDateLabel: "2023-08-19", routeLabel: "컴구, 데통>컴네>등등", contentLabel: "고렇게만 듣자!")
+
